refactor(_cadqueryapi): split into a package, extract _placement and _patches
Step 1 of 3 in the _cadqueryapi.py teardown.

* New package layout under bluemira/codes/_cadqueryapi/:
  - _aliases.py    (apiVertex/Edge/… + numerical tolerances)
  - _placement.py  (Vector/Placement/Plane adapters + make_placement/make_plane/…)
  - _patches.py    (cq.Wire/Face/Solid monkey-patches; side-effect-only)
  - _core.py       (everything else, ~3000 lines — split further in PR 2/3)
  - __init__.py    (explicit re-exports + __all__ aggregation)

* __init__.__getattr__ falls through to submodules before raising
  NotImplementedError, preserving the old single-file behaviour where every
  module-level name (including underscore-prefixed helpers) was reachable via
  getattr — depended on by the _geometryapi dispatcher and a few tests.

* Error types (FreeCADError, InvalidCADInputsError, GeometryError) re-exported
  from the package root via PEP 484 ``import X as X`` so cadapi.FreeCADError
  keeps resolving for backend-agnostic exception catching.

Public API surface unchanged (verified via __all__ aggregation + test suite).
