+++ pyproject.toml
@@ -331,12 +331,12 @@ ignore-names = [
 "bluemira/__init__.py" = ["TID252"]
 "bluemira/codes/__init__.py" = ["E402"]
 "bluemira/codes/process/model_mapping.py" = ["PLR6301", "DOC201"]
-# _cadqueryapi.py is an experimental backend mirroring _freecadapi's
+# _cadqueryapi/* is an experimental backend mirroring _freecadapi's
 # surface; several patterns are either inherited from FreeCAD's API
 # (CapWords-violating names, 1-line docstrings without Returns/Raises)
 # or live in adapter shims where args are intentionally unused to match
 # the other backend's signature.
-"bluemira/codes/_cadqueryapi.py" = [
+"bluemira/codes/_cadqueryapi/*.py" = [
     "DOC201",
     "DOC501",
     "ARG001",
